refactor(j-s): Extract indictment reviewer selector into a new component
